+++ src/main/java/io/github/mortuusars/sootychimneys/data/builder/SootScrapingRecipeBuilder.java
@@ -12,19 +12,22 @@ import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.registries.ForgeRegistries;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Consumer;
 
 public class SootScrapingRecipeBuilder {
-    private final List<ChanceResult> results = new ArrayList<>(4);
-    private final Ingredient input;
+    private final List<ChanceResult> results = new ArrayList<>(3);
+    private final Ingredient chimney;
     private final Ingredient tool;
 
-    public SootScrapingRecipeBuilder(Ingredient input, Ingredient tool, ItemLike mainResult, int count, float chance) {
-        this.input = input;
+    public SootScrapingRecipeBuilder(Ingredient chimney, Ingredient tool, ItemLike mainResult, int count, float chance) {
+        this.chimney = chimney;
         this.tool = tool;
         this.results.add(new ChanceResult(new ItemStack(mainResult.asItem(), count), chance));
     }
@@ -48,14 +51,14 @@ public class SootScrapingRecipeBuilder {
     }
 
     public void build(Consumer<FinishedRecipe> consumerIn) {
-        ResourceLocation location = ForgeRegistries.ITEMS.getKey(this.input.getItems()[0].getItem());
-        if (location.toString().equals("minecraft:barrier"))
-            throw new IllegalStateException("Cannot get path from the ingredient: '" + input.toJson() + "'. Provide custom save path.");
+        ResourceLocation location = ForgeRegistries.ITEMS.getKey(this.chimney.getItems()[0].getItem());
+        if (Objects.requireNonNull(location).toString().equals("minecraft:barrier"))
+            throw new IllegalStateException("Cannot get path from the ingredient: '" + chimney.toJson() + "'. Provide custom save path.");
         this.build(consumerIn, location.getPath());
     }
 
     public void build(Consumer<FinishedRecipe> consumerIn, String save) {
-        ResourceLocation resourcelocation = ForgeRegistries.ITEMS.getKey(this.input.getItems()[0].getItem());
+        ResourceLocation resourcelocation = ForgeRegistries.ITEMS.getKey(this.chimney.getItems()[0].getItem());
         ResourceLocation saveLocation = new ResourceLocation(SootyChimneys.MOD_ID, "soot_scraping/" + save);
         if (saveLocation.equals(resourcelocation)) {
             throw new IllegalStateException("Soot Scraping Recipe '" + save + "' should remove its 'save' argument as it's the same as the default.");
@@ -66,27 +69,26 @@ public class SootScrapingRecipeBuilder {
 
     public void build(Consumer<FinishedRecipe> consumerIn, ResourceLocation id) {
         // Sort outputs with larger chance first:
-        results.sort((result1, result2) -> result1.getChance() > result2.getChance() ? -1 : 1);
-        consumerIn.accept(new SootScrapingRecipeBuilder.Result(id, this.input, this.tool, this.results));
+        results.sort(Comparator.comparingDouble(ChanceResult::getChance));
+        consumerIn.accept(new SootScrapingRecipeBuilder.Result(id, this.chimney, this.tool, this.results));
     }
 
     public static class Result implements FinishedRecipe {
-
         private final ResourceLocation id;
-        private final Ingredient ingredient;
+        private final Ingredient chimney;
         private final Ingredient tool;
         private final List<ChanceResult> results;
 
         public Result(ResourceLocation id, Ingredient ingredient, Ingredient tool, List<ChanceResult> results) {
             this.id = id;
-            this.ingredient = ingredient;
+            this.chimney = ingredient;
             this.tool = tool;
             this.results = results;
         }
 
         @Override
         public void serializeRecipeData(JsonObject json) {
-            json.add("ingredient", this.ingredient.toJson());
+            json.add("chimney", this.chimney.toJson());
             json.add("tool", this.tool.toJson());
 
             JsonArray arrayResults = new JsonArray();
@@ -97,12 +99,12 @@ public class SootScrapingRecipeBuilder {
         }
 
         @Override
-        public ResourceLocation getId() {
+        public @NotNull ResourceLocation getId() {
             return this.id;
         }
 
         @Override
-        public RecipeSerializer<?> getType() {
+        public @NotNull RecipeSerializer<?> getType() {
             return ModRecipeSerializers.SOOT_SCRAPING.get();
         }