+++ src/main/java/io/github/mortuusars/sootychimneys/integration/jei/category/SootScrapingRecipeCategory.java
@@ -1,12 +1,11 @@
 package io.github.mortuusars.sootychimneys.integration.jei.category;
 
-import com.google.common.collect.ImmutableList;
 import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.mortuusars.sootychimneys.SootyChimneys;
-import io.github.mortuusars.sootychimneys.config.Config;
 import io.github.mortuusars.sootychimneys.integration.jei.JeiRecipeTypes;
-import io.github.mortuusars.sootychimneys.integration.jei.recipe.SootScrapingJeiRecipe;
 import io.github.mortuusars.sootychimneys.integration.jei.renderer.ScalableItemStackRenderer;
+import io.github.mortuusars.sootychimneys.recipe.SootScrapingRecipe;
+import io.github.mortuusars.sootychimneys.recipe.ingredient.ChanceResult;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -17,29 +16,23 @@ import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
+import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.NotNull;
 
-import java.util.Collections;
 import java.util.List;
 
-@SuppressWarnings("removal")
-public class SootScrapingRecipeCategory implements IRecipeCategory<SootScrapingJeiRecipe> {
-//    public static final ResourceLocation UID = SootyChimneys.resource("soot_scraping");
-
+public class SootScrapingRecipeCategory implements IRecipeCategory<SootScrapingRecipe> {
     public static final int BG_WIDTH = 153;
     public static final int BG_HEIGHT = 65;
-
-    private final List<Component> BYPRODUCT_ITEMS_INFO = ImmutableList.of(
-            Component.translatable("jei." + SootyChimneys.MOD_ID + ".category.soot_scraping.soot_items_info"));
 
     private final Component title;
     private final IDrawable background;
     private final IDrawableStatic icon;
-    private final IDrawableStatic byproductInfo;
-
-    private final int byproductXPos = 70;
-    private final int byproductYPos = 37;
+    private final IDrawableStatic dust;
+    private final IDrawableStatic slot;
+    private final IDrawableStatic chanceSlot;
 
     public SootScrapingRecipeCategory(IGuiHelper helper) {
         title = Component.translatable("jei." + SootyChimneys.MOD_ID + ".category.soot_scraping");
@@ -51,70 +44,90 @@ public class SootScrapingRecipeCategory implements IRecipeCategory<SootScrapingJ
                 .setTextureSize(16, 16)
                 .build();
 
-        byproductInfo = helper.createDrawable(texture, 153, 0, 14, 25);
+        dust = helper.createDrawable(texture, 155, 0, 10, 12);
+        slot = helper.createDrawable(texture, 0, 65, 18, 18);
+        chanceSlot = helper.createDrawable(texture, 18, 65, 18, 18);
     }
 
     @Override
-    public void setRecipe(IRecipeLayoutBuilder builder, SootScrapingJeiRecipe recipe, IFocusGroup focuses) {
-        builder.addSlot(RecipeIngredientRole.INPUT, 10, 18)
+    public void setRecipe(IRecipeLayoutBuilder builder, SootScrapingRecipe recipe, @NotNull IFocusGroup focuses) {
+        builder.addSlot(RecipeIngredientRole.INPUT, 9, 18)
                 .setCustomRenderer(VanillaTypes.ITEM_STACK, new ScalableItemStackRenderer(2.5f))
-                .addItemStack(recipe.getIngredientChimney())
+                .addIngredients(recipe.getIngredientChimney())
                 .setSlotName("DirtyChimney");
 
         builder.addSlot(RecipeIngredientRole.CATALYST, 44, 1)
-                        .addIngredients(VanillaTypes.ITEM_STACK, recipe.getTools());
+                .addIngredients(recipe.getTool())
+                .setSlotName("Tool");
 
-        builder.addSlot(RecipeIngredientRole.OUTPUT, 103, 18)
+        builder.addSlot(RecipeIngredientRole.OUTPUT, 104, 18)
                 .setCustomRenderer(VanillaTypes.ITEM_STACK, new ScalableItemStackRenderer(2.5f))
                 .addItemStack(recipe.getResultChimney())
                 .setSlotName("CleanChimney");
-    }
 
-    @Override
-    public void draw(SootScrapingJeiRecipe recipe, IRecipeSlotsView recipeSlotsView, PoseStack stack, double mouseX, double mouseY) {
-        if (Config.DISPLAY_JEI_SCRAPING_BYPRODUCTS_INFO.get()) {
-            byproductInfo.draw(stack, byproductXPos, byproductYPos);
+        List<ChanceResult> results = recipe.getResults()
+                .stream()
+                .filter(result -> !result.getStack().isEmpty() && result.getChance() > 0)
+                .toList();
+
+        int slotX = 51 + (27 - ((results.size() * 18) / 2));
+
+        for (int i = 0; i < Math.min(results.size(), SootScrapingRecipe.MAX_RESULTS); i++) {
+            ChanceResult result = results.get(i);
+            builder.addSlot(RecipeIngredientRole.OUTPUT, slotX + (18 * i) - 1, 47)
+                    .setSlotName("Result" + i + 1)
+                    .addItemStack(result.getStack())
+                    .addTooltipCallback((recipeSlotView, tooltip) -> {
+                        if (result.getChance() < 1.0f) {
+                            float chance = result.getChance() * 100;
+                            String chanceString = chance < 1 ? "<1" : Integer.toString((int) chance);
+                            tooltip.add(Component.translatable("jei.sootychimneys.chance", chanceString).withStyle(ChatFormatting.GOLD));
+                        }
+                    });
         }
     }
 
     @Override
-    public List<Component> getTooltipStrings(SootScrapingJeiRecipe recipe, IRecipeSlotsView recipeSlotsView, double mouseX, double mouseY) {
-        if (Config.DISPLAY_JEI_SCRAPING_BYPRODUCTS_INFO.get()) {
-            if ((mouseX >= byproductXPos && mouseX < byproductXPos + byproductInfo.getWidth())
-                    && (mouseY >= byproductYPos && mouseY < byproductYPos + byproductInfo.getHeight()))
-                return BYPRODUCT_ITEMS_INFO;
-        }
+    public void draw(@NotNull SootScrapingRecipe recipe, @NotNull IRecipeSlotsView recipeSlotsView, @NotNull PoseStack stack, double mouseX, double mouseY) {
+        List<ChanceResult> results = recipe.getResults()
+                .stream()
+                .filter(result -> !result.getStack().isEmpty() && result.getChance() > 0)
+                .toList();
 
-        return Collections.emptyList();
+        if (results.isEmpty())
+            return;
+
+        dust.draw(stack, 72, 34);
+
+        int slotX = 49 + (27 - ((results.size() * 18) / 2));
+
+        for (int i = 0; i < results.size(); i++) {
+            ChanceResult result = results.get(i);
+            if (result.getChance() >= 1.0f)
+                slot.draw(stack, slotX + (18 * i), 46);
+            else
+                chanceSlot.draw(stack, slotX + (18 * i), 46);
+        }
     }
 
     @Override
-    public Component getTitle() {
+    public @NotNull Component getTitle() {
         return title;
     }
 
     @Override
-    public IDrawable getBackground() {
+    public @NotNull IDrawable getBackground() {
         return background;
     }
 
     @Override
-    public IDrawable getIcon() {
+    public @NotNull IDrawable getIcon() {
         return icon;
     }
 
-//    @Override
-//    public ResourceLocation getUid() {
-//        return UID;
-//    }
-//
-//    @Override
-//    public Class<? extends SootScrapingJeiRecipe> getRecipeClass() {
-//        return this.getRecipeType().getRecipeClass();
-//    }
-
     @Override
-    public RecipeType<SootScrapingJeiRecipe> getRecipeType() {
+    public @NotNull RecipeType<SootScrapingRecipe> getRecipeType() {
         return JeiRecipeTypes.SOOT_SCRAPING;
     }
 }
+