+++ src/main/java/io/github/mortuusars/sootychimneys/recipe/ingredient/ChanceResult.java
@@ -8,6 +8,7 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.core.Registry;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.util.RandomSource;
 import net.minecraft.world.item.ItemStack;
 
 import java.util.Optional;
@@ -18,11 +19,10 @@ import java.util.Random;
  */
 @SuppressWarnings({"deprecation", "ClassCanBeRecord", "Convert2MethodRef"})
 public class ChanceResult {
-
     public static final Codec<ChanceResult> CODEC = RecordCodecBuilder.create(instance -> instance.group(
                     // Not using ItemStack.CODEC here. This way json is more intuitive, without one more object.
                     Registry.ITEM.byNameCodec().fieldOf("item").forGetter(chanceResult -> chanceResult.stack.getItem()),
-                    Codec.INT.optionalFieldOf("count", 1).forGetter(chanceResult -> chanceResult.stack.getCount()),
+                    Codec.INT.optionalFieldOf("Count", 1).forGetter(chanceResult -> chanceResult.stack.getCount()),
                     CompoundTag.CODEC.optionalFieldOf("tag").forGetter(chanceResult -> Optional.ofNullable(chanceResult.stack.getTag())),
                     Codec.FLOAT.optionalFieldOf("chance", 1F).forGetter(ChanceResult::getChance))
             .apply(instance, (item, count, compoundTag, chance) -> {
@@ -53,7 +53,7 @@ public class ChanceResult {
         return chance;
     }
 
-    public ItemStack rollOutput(Random rand) {
+    public ItemStack rollOutput(RandomSource rand) {
         int outputAmount = stack.getCount();
         for (int roll = 0; roll < stack.getCount(); roll++)
             if (rand.nextFloat() > chance)